Refactor clasp logs
